backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==64.0.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
feedparser==6.0.12
idna==3.11
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
sgmllib3k==1.0.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.0
google-generativeai==0.8.6
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
python-dotenv==1.2.1
PyPDF2==3.0.1